@ fix(core): signal, animation, utf8, color, i18n and widget-tree
- signal.h: emit() is re-entrancy safe (slots may destroy the Signal); emitCopy
  for one-shot animations; recursive A->B->A emit detected.
- animation: stop/pause/dtors stop underlay AnimatedFloat (activeCount never
  leaks); onFinished uses emitCopy; play() re-init on re-target; keyframe
  anims count into activeCount; zero-dt keyframes skipped.
- utf8: nextPos clamps at string end.
- color: toARGB/toABGR cast to uint32 before shift.
- log: unregistered categories enabled; NDEBUG gates Debug/Info only.
- widget/window: focus clearing on reparent (UAF), dispatch snapshot made
  local, onPaint dirty-rect accumulation preserved, menu/context/modal open
  states cleaned in destructor, hitTest coordinate order.
- i18n: CLDR plural rules for fa/hi/bn/gu/kn/ur/zu/pa/ne/si/is/mk/pt/lv.
- layout: removeChild clears stale stretch entries; geometry changes invalidate.
- app/main: centered label via leading+trailing stretch.

+++ include/ltgui/color.h
@@ -25,7 +25,11 @@ struct Color {
   }
   bool operator!=(const Color &o) const { return !(*this == o); }
 
-  uint32_t toARGB() const { return (a << 24) | (r << 16) | (g << 8) | b; }
+  uint32_t toARGB() const {
+    return (static_cast<uint32_t>(a) << 24) |
+           (static_cast<uint32_t>(r) << 16) | (static_cast<uint32_t>(g) << 8) |
+           static_cast<uint32_t>(b);
+  }
 
   // Linear interpolation per channel (alpha included). t=0 -> a, t=1 -> b.
   static Color lerp(const Color &a, const Color &b, float t) {
@@ -40,7 +44,11 @@ struct Color {
   // DXGI_FORMAT_R8G8B8A8_UNORM and GL_RGBA+GL_UNSIGNED_BYTE both
   // read bytes at increasing offsets as [R][G][B][A], which in LE
   // corresponds to the uint32_t layout 0xAABBGGRR.
-  uint32_t toABGR() const { return (a << 24) | (b << 16) | (g << 8) | r; }
+  uint32_t toABGR() const {
+    return (static_cast<uint32_t>(a) << 24) |
+           (static_cast<uint32_t>(b) << 16) | (static_cast<uint32_t>(g) << 8) |
+           static_cast<uint32_t>(r);
+  }
 
   static const Color Transparent;
   static const Color Black;
